requirements.txt
py-cord[voice]
dislash.py
jishaku
EZPaginator
pathlib
mysql-connector
colored